# spring-server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.0.11
servers:
- url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /user:
    get:
      description: Returns list of users
      operationId: getUserList
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
      summary: Get list of all users
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
    post:
      description: Add a new user to the user
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
        description: Create a new user in the user
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new user to the user
      tags:
      - user
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: user
    put:
      description: Update an existing user by Id
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
        description: Update an existent user in the user
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "405":
          description: Validation exception
      summary: Update an existing user
      tags:
      - user
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: user
  /user/{userId}:
    delete:
      description: delete a user
      operationId: deleteUser
      parameters:
      - description: User id to delete
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: successful operation
        "400":
          description: Invalid user value
      summary: Deletes a user
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
    get:
      description: Returns a single user
      operationId: getUser
      parameters:
      - description: ID of user to return
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
      summary: Find user by ID
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
...
components:
...
  schemas:
...
    User:
      example:
        firstName: John
        lastName: James
        password: "12345"
        phone: "12345"
        id: 1
        email: john@xyz.com
        username: theUser
      properties:
        id:
          example: 1
          format: int64
          type: integer
        username:
          example: theUser
          type: string
        firstName:
          example: John
          type: string
        lastName:
          example: James
          type: string
        email:
          example: john@xyz.com
          type: string
        password:
          example: "12345"
          type: string
        phone:
          example: "12345"
          type: string
      type: object
      xml:
        name: user
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey